refactor: remove duplicate logic
