+++ src/components/layout/layout.tsx
@@ -3,53 +3,43 @@ import { Header } from '../header';
 import { Main } from '../main';
 import { fetchData } from '../../services/api.ts';
 import { BASE_URL, SEARCH_PARAM } from '../../constants/api.ts';
-import { getItemFromLocalStorage, setItemToLocalStorage } from '../../utils/utils.ts';
+import { getItemFromLocalStorage, setItemToLocalStorage } from '../../utils';
 import { Data } from '../../types';
 import { Loader } from '../loader';
 import { Error } from '../error';
 
 interface LayoutState {
+  data: Data | { info: Record<string, unknown>; results: [] };
   loading: boolean;
   error: boolean;
   searchQuery: string;
-  data: Data | { info: Record<string, unknown>; results: [] };
 }
 
 export class Layout extends Component<Record<string, never>, LayoutState> {
   state = {
+    data: { info: {}, results: [] },
     loading: false,
     error: false,
     searchQuery: getItemFromLocalStorage('searchQuery'),
-    data: { info: {}, results: [] },
   };
 
   componentDidMount() {
     this.handleSearch();
   }
 
-  setData = (data: Data) => {
+  updateState = (newState: Partial<LayoutState>) => {
     this.setState((prevState) => ({
       ...prevState,
-      data,
+      ...newState,
     }));
   };
 
   handleInputChange = (e: ChangeEvent<HTMLInputElement>) => {
-    this.setState((prevState) => ({
-      ...prevState,
-      searchQuery: e.target.value,
-    }));
+    this.updateState({ searchQuery: e.target.value });
   };
 
   handleSearch = () => {
-    const { searchQuery } = this.state;
-    const savedSearchQuery = getItemFromLocalStorage('searchQuery');
-
-    if (savedSearchQuery) {
-      this.fetchByQuery(savedSearchQuery);
-      return;
-    }
-
+    const searchQuery = getItemFromLocalStorage('searchQuery') || this.state.searchQuery;
     this.fetchByQuery(searchQuery);
   };
 
@@ -60,34 +50,20 @@ export class Layout extends Component<Record<string, never>, LayoutState> {
   };
 
   handleRefresh = () => {
-    this.setState((prevState) => ({
-      ...prevState,
-      searchQuery: '',
-      error: false,
-    }));
+    this.updateState({ searchQuery: '', error: false });
     setItemToLocalStorage('searchQuery', '');
     this.fetchByQuery('');
   };
 
   fetchByQuery = async (searchQuery: string) => {
     try {
-      this.setState((prevState) => ({
-        ...prevState,
-        loading: true,
-        error: false,
-      }));
+      this.updateState({ loading: true, error: false });
       const data = await fetchData<Data>(`${BASE_URL}/?${SEARCH_PARAM}=${searchQuery}`);
-      this.setData(data);
+      this.updateState({ data });
     } catch (error) {
-      this.setState((prevState) => ({
-        ...prevState,
-        error: true,
-      }));
+      this.updateState({ error: true });
     } finally {
-      this.setState((prevState) => ({
-        ...prevState,
-        loading: false,
-      }));
+      this.updateState({ loading: false });
     }
   };
