+++ src/styles/main.css
@@ -4,7 +4,6 @@
 html,
 body {
   min-height: 100vh;
-  /*max-width: 1200px;*/
   margin: 0 auto;
   font-family: 'DM Sans', sans-serif;
   color: #323232;
